Класс Node вынес в отдельный файл. Исправил ошибку в методе InMemoryTaskManager.
